Back-End/login/requirements.txt
asgiref==3.8.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==44.0.0
Django==5.1.3
django-cors-headers==4.6.0
django-oauth-toolkit==3.0.1
django-ratelimit==4.1.0
djangorestframework==3.15.2
idna==3.10
jwcrypto==1.5.6
oauthlib==3.2.2
psycopg==3.2.5
psycopg-binary==3.2.5
pycparser==2.22
requests==2.32.3
sqlparse==0.5.2
typing_extensions==4.12.2
urllib3==2.2.3
djangorestframework-simplejwt==5.4.0
    # via -r Back-End/requirements.in
drf-yasg==1.21.8
    # via -r Back-End/requirements.in
gunicorn==23.0.0
pyotp==2.9.0
django-encrypted-model-fields==0.6.5
